+++ testData/org/elixir_lang/psi/impl/declarations/issue_2491/anonymous_function_with_stab.ex
@@ -1,0 +1,1 @@
+fn <caret>p -> end